backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.6.0

# Google AI
google-genai==1.0.0

# PDF Processing
pymupdf==1.25.0
pdfplumber==0.11.4

# Vector Store
faiss-cpu==1.9.0
numpy==2.2.0

# Async Support
aiofiles==24.1.0
httpx==0.28.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Utilities
tiktoken==0.8.0
tenacity==9.0.0
